Win32 모더나이제이션 후속 정리
- COM STA 초기화를 main()으로 일원화 (file_dialog의 호출별 CoInit/CoUninit 제거)
- windows crate 노출 상수의 직접 정의 제거 (CB_*/LB_*/EM_*/CFM_*/SS_*/EVENT_* 등 30여 개)
  - Win32_UI_Controls_RichEdit, Win32_System_Ole, Win32_System_SystemServices feature 추가
  - backlog.rs의 수동 CHARFORMAT2W 정의 제거 (windows crate 정식 타입 사용)
- 클립보드 HANDLE→HGLOBAL 변환에서 transmute 제거 (newtype .0 명시 변환)
- TrayIcon szTip 설정을 set_sz_tip 헬퍼로 추출 (create/update_tooltip 중복 제거)
- 마이크로: shared_client의 redundant_closure 제거, RESPONSE_STORAGE의 Arc 제거

빌드: i686-pc-windows-msvc 0 errors / 31 warnings (모두 기존 dead-code).

+++ src/dialogs/settings/mod.rs
@@ -18,7 +18,7 @@ use windows::{
 };
 
 use crate::config::{Config, TextAlign};
-use crate::constants::{CB_ADDSTRING, CB_RESETCONTENT, CB_SETCURSEL, TBM_GETPOS_VAL, TCN_SELCHANGE};
+use crate::constants::TBM_GETPOS_VAL;
 use crate::impl_dialog;
 use crate::translation::{TranslationEngine, lang_utils};
 use crate::util::to_wide;
@@ -740,7 +740,6 @@ impl SettingsDialog {
             // 초기 항목 채우기
             {
                 let cfg = self.config.borrow();
-                use crate::constants::LB_ADDSTRING;
                 for k in &cfg.translation.deepl_keys {
                     let kw = to_wide(k);
                     let _ = SendMessageW(listbox, LB_ADDSTRING, Some(WPARAM(0)), Some(LPARAM(kw.as_ptr() as isize)));
@@ -917,7 +916,7 @@ impl SettingsDialog {
                 // SAFETY: lparam points to a valid NMHDR struct from the system.
                 unsafe {
                     let nmhdr = &*(lparam.0 as *const NMHDR);
-                    if nmhdr.code == TCN_SELCHANGE as u32 {
+                    if nmhdr.code == TCN_SELCHANGE {
                         if let Ok(tab_hwnd) = GetDlgItem(Some(self.hwnd), ctrl_id::TAB_CONTROL as i32) {
                             let sel = SendMessageW(
                                 tab_hwnd,
